update wallet code
